+++ CHANGELOG.md
@@ -14,7 +14,11 @@
   - in addition the text of the headline will be selected
 - if a note was created by hitting `Return` in the note search line edit now
   the line edit will be cleared so all notes will be viewed again
-  
+- <kbd>Ctrl</kbd> + <kbd>Space</kbd> now tries to solve simple equations like
+ `(4+5)*3=` in the current line in front of the cursor 
+  (for [Issue #152](https://github.com/pbek/QOwnNotes/issues/152))
+  - the result will be put into the note if the cursor is after the `=` 
+
 ## 16.05.1
 - fixed the automatic image resizing in the preview under Windows
   (for [Issue #154](https://github.com/pbek/QOwnNotes/issues/154))